requirements.txt
pytest==6.2.5
Faker==8.12.1
mypy==0.910
flake8==3.9.2
sqlalchemy==1.4.25
flask==2.0.2
requests==2.26.0
psycopg2-binary==2.9.1
alembic==1.7.4
arrow==1.2.0
PyJWT==1.7.1